updated file names/naming to match sst-elements conventions
